corrected some SDL-Header includes; some fixes on the linux side & Debug code
git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@1124 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/ui/aui_common/aui_hypertextbase.cpp
@@ -153,7 +153,7 @@ AUI_ERRCODE aui_HyperTextBase::SetHyperText
 	if (hyperText)
     {
         strncpy(m_hyperText, hyperText, std::min(maxlen, m_hyperMaxLen));
-        m_hyperCurLen = std::min(m_hyperMaxLen, strlen(m_hyperText));
+        m_hyperCurLen = std::min((size_t) m_hyperMaxLen, strlen(m_hyperText));
     	return AddHyperStatics(NULL);
     }
     else
